refactor: entity id 타입 integer -> long 변환에 따른 변경

+++ Backend/triplet/src/main/java/com/ssafy/triplet/exchange/repository/ExchangeResultsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.triplet.exchange.domain.ExchangeRecord;
 
-public interface ExchangeResultsRepository extends JpaRepository<ExchangeRecord, Integer> {
+public interface ExchangeResultsRepository extends JpaRepository<ExchangeRecord, Long> {
     List<ExchangeRecord> findByNameAndPhoneNumAndBirth(String name, String phoneNum, String birth);
 }
